services/data-ingestor/requirements.txt
httpx==0.27.*
minio==7.*
pika==1.3.*
python-dotenv==1.0.*
